moment instlairan namestena pocetna stranica

+++ database/migrations/2020_06_19_145914_create_uceniks_table.php
@@ -18,7 +18,7 @@ class CreateUceniksTable extends Migration
             $table->string('ime_prezime');
             $table->date('datum_rodjenja');
             $table->decimal('broj_bodova', 11, 2);
-            $table->integer('jmbg');
+            $table->bigInteger('jmbg');
             $table->integer('pol');
             $table->unsignedBigInteger('odeljenje_id')->nullable();
             $table->timestamps();
